requirements.txt
Flask==3.1.1
discord.py==2.6.2
python-dotenv==1.1.1
aiosqlite==0.21.0
aiohttp==3.12.15
